# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "StudentPlanner.UI",
    "version": "1.0"
  },
  "paths": {
...
    "/account/register": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Registers a new user account.",
        "description": "This endpoint creates a new user account using the provided registration details.\r\nIf the email address is already in use, the operation fails.",
        "requestBody": {
          "description": "The user registration details.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "If the provided password does not meet validation rules.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "If an account with the given email already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "204": {
            "description": "User registered successfully."
          }
        }
      }
    },
    "/account/login": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Authenticates a user using email and password.",
        "description": "If authentication succeeds, a JSON Web Token (JWT) is issued and returned\r\nto the client. The token must be included in subsequent requests to\r\nauthenticate the user.",
        "requestBody": {
          "description": "The login credentials.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "If the provided credentials are invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/account/logout": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Logs out the currently authenticated user.",
        "description": "This endpoint terminates the current authentication session.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "204": {
            "description": "Logout successful."
          },
          "401": {
            "description": "If the request is not authenticated."
          }
        }
      }
    },
    "/account/delete-account": {
      "delete": {
        "tags": [
          "Account"
        ],
        "summary": "Deletes the authenticated user's account.",
        "description": "Only users with the `User` role may delete their own accounts.\r\nThe operation signs the user out before account deletion.",
        "parameters": [
          {
            "name": "userEmail",
            "in": "query",
            "description": "The email address of the account to delete.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "If the user does not have permission to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "204": {
            "description": "Account deleted successfully."
          }
        }
      }
    },
    "/admin/create-manager": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Creates a new manager account.",
        "description": "This endpoint creates a new user account and assigns the Manager role.\r\nOnly administrators are authorized to perform this operation.",
        "requestBody": {
          "description": "The manager account creation details.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "If the manager already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Manager account created successfully."
          },
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/admin/delete-manager": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Deletes an existing manager account.",
        "description": "Only administrators are allowed to delete manager accounts.",
        "parameters": [
          {
            "name": "ManagerEmail",
            "in": "query",
            "description": "The email address of the manager to delete.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "If the manager does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "If the email address is not provided."
          },
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "204": {
            "description": "Manager deleted successfully."
          }
        }
      }
    },
    "/admin": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Updates the status of an event request.",
        "description": "This endpoint allows an administrator to review an existing event request\r\nand update its status. It is used to approve or reject requests submitted\r\nby managers for the creation, modification, or deletion of academic events.",
        "requestBody": {
          "description": "The request containing the identifier of the event request and the new status\r\nto be applied. The status must be set to either `Approved` or `Rejected`.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The event request status was updated successfully."
          },
          "400": {
            "description": "If the provided request data is invalid or the status value is not supported.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "If the authenticated user does not have administrative privileges.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "If the specified event request does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AcademicEventResponse": {
        "required": [
          "details",
          "eventId",
          "faculty"
        ],
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "faculty": {
            "$ref": "#/components/schemas/FacultyDTO"
          },
          "details": {
            "$ref": "#/components/schemas/EventDetailsOutputDTO"
          }
        },
        "additionalProperties": false
      },
...
      "CreateUserRequest": {
        "required": [
          "confirmPassword",
          "email",
          "firstName",
          "password",
          "surname"
        ],
        "type": "object",
        "properties": {
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          },
          "confirmPassword": {
            "minLength": 1,
            "type": "string"
          },
          "firstName": {
            "minLength": 1,
            "type": "string"
          },
          "surname": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "EventDetailsOutputDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FacultyDTO": {
        "required": [
          "displayName",
          "facultyId",
          "name"
        ],
        "type": "object",
        "properties": {
          "facultyId": {
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "displayName": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "LoginResponse": {
        "required": [
          "accessToken",
          "email",
          "expiresAt",
          "firstName",
          "surname"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "minLength": 1,
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "firstName": {
            "minLength": 1,
            "type": "string"
          },
          "surname": {
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "personalEvents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PersonalEventResponse"
            },
            "nullable": true
          },
          "academicEvents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AcademicEventResponse"
            },
            "nullable": true
          },
          "usosEvents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UsosEventResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PersonalEventResponse": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "details": {
            "$ref": "#/components/schemas/EventDetailsOutputDTO"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "UpdateEventRequest": {
        "required": [
          "requestId",
          "requestStatus"
        ],
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "requestStatus": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "UsosEventResponse": {
        "required": [
          "eventDetailsOutputDTO"
        ],
        "type": "object",
        "properties": {
          "eventDetailsOutputDTO": {
            "$ref": "#/components/schemas/EventDetailsOutputDTO"
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}